+++ CascableCore.xcframework/ios-arm64/CascableCore.framework/Headers/CBLPlatformDefines.h
@@ -16,12 +16,12 @@
 #define PLATFORM_IMAGE NSImage
 
 @interface NSImage (CBLPlatform)
-+(NSImage *)imageWithCGImage:(CGImageRef)cgImage;
++(NSImage * _Nonnull)imageWithCGImage:(CGImageRef _Nonnull)cgImage;
 @end
 
 #endif
 
 /// Returns a CascableCore localised string from the given table. Typically this shouldn't be used - use the various
 /// localised value properties in the framework instead.
-extern NSString * CBLCoreLocalizedString(NSString *key, NSString *table);
+extern NSString * _Nonnull CBLCoreLocalizedString(NSString * _Nonnull key, NSString * _Nonnull table);
 #define CBLCoreLocalizedStringWithFormat(key, table, ...) [NSString stringWithFormat:CBLCoreLocalizedString(key, table), __VA_ARGS__]